Move Landing page to separate folder and add to Storybook
